+++ tests/testthat/test-req-body.R
@@ -47,12 +47,20 @@ test_that("can send named list as json/form/multipart", {
   expect_equal(json$form, list(a = "1", b = "2"))
 })
 
+test_that("body data preserves data frame", {
+  req1 <- request_test() %>% req_body_form(mtcars)
+  expect_equal(req1$body$data, mtcars)
+})
+
 test_that("can modify body data", {
   req1 <- request_test() %>% req_body_form(list(a = 1))
-  req2 <- req1 %>% req_body_form(list(b = 2))
-
   expect_equal(req1$body$data, list(a = 1))
+
+  req2 <- req1 %>% req_body_form(list(b = 2))
   expect_equal(req2$body$data, list(a = 1, b = 2))
+
+  req3 <- req1 %>% req_body_form(list(a = 3, a = 4))
+  expect_equal(req3$body$data, list(a = 3, a = 4))
 })
 
 test_that("can upload file with multipart", {