+++ tests/testthat/test-utils.R
@@ -1,12 +1,19 @@
 test_that("modify list adds, removes, and overrides", {
   x <- list(x = 1)
-  expect_equal(modify_list(x), x)
-  expect_equal(modify_list(x, x = NULL), list())
-  expect_equal(modify_list(x, x = 2), list(x = 2))
-  expect_equal(modify_list(x, y = 3), list(x = 1, y = 3))
-  expect_equal(modify_list(NULL, x = 2), list(x = 2))
+  expect_equal(modify_list_dots(x), x)
+  expect_equal(modify_list_dots(x, x = NULL), list())
+  expect_equal(modify_list_dots(x, x = 2), list(x = 2))
+  expect_equal(modify_list_dots(x, y = 3), list(x = 1, y = 3))
+  expect_equal(modify_list_dots(NULL, x = 2), list(x = 2))
 
-  expect_snapshot(modify_list(x, a = 1, 2), error = TRUE)
+  expect_snapshot(modify_list_dots(x, a = 1, 2), error = TRUE)
+})
+
+test_that("replacement affects all components with name", {
+  x <- list(a = 1, a = 2)
+  expect_equal(modify_list_dots(x, a = NULL), list())
+  expect_equal(modify_list_dots(x, a = 3), list(a = 3))
+  expect_equal(modify_list_dots(x, a = 3, a = 4), list(a = 3, a =4))
 })
 
 test_that("can check arg types", {