+++ R/req-options.R
@@ -84,6 +84,46 @@ req_timeout <- function(req, seconds) {
   req_options(req, timeout_ms = seconds * 1000)
 }
 
+
+#' Use a proxy for a request
+#'
+#' @inheritParams req_perform
+#' @param url,port Location of proxy.
+#' @param username,password Login details for proxy, if needed.
+#' @param auth Type of HTTP authentication to use. Should be one of the
+#'   following: `basic`, digest, digest_ie, gssnegotiate, ntlm, any.
+#' @examples
+#' # Proxy from https://www.proxynova.com/proxy-server-list/
+#' \dontrun{
+#' request("http://hadley.nz") %>%
+#'   req_proxy("20.116.130.70", 3128) %>%
+#'   req_perform()
+#' }
+#' @export
+req_proxy <- function(req, url, port = NULL, username = NULL, password = NULL, auth = "basic") {
+
+  if (!is.null(username) || !is.null(password)) {
+    proxyuserpwd <- paste0(username, ":", password)
+  } else {
+    proxyuserpwd <- NULL
+  }
+
+  if (!is.null(port)) {
+    if (!is_integerish(port)) {
+      abort("`port` must be a number")
+    }
+  }
+
+  req_options(
+    req,
+    proxy = url,
+    proxyport = port,
+    proxyuserpwd = proxyuserpwd,
+    proxyauth = auth_flags(auth)
+  )
+}
+
+
 #' Show extra output when request is performed
 #'
 #' @description
@@ -142,6 +182,7 @@ req_verbose <- function(req,
   req_options(req, debugfunction = debug, verbose = TRUE)
 }
 
+
 # helpers -----------------------------------------------------------------
 
 verbose_message <- function(prefix, x) {
@@ -169,4 +210,17 @@ verbose_header <- function(prefix, x, redact = TRUE) {
       cli::cat_line(prefix, line)
     }
   }
+}
+
+auth_flags <- function(x = "basic") {
+  constants <- c(
+    basic = 1,
+    digest = 2,
+    gssnegotiate = 4,
+    ntlm = 8,
+    digest_ie = 16,
+    any = -17
+  )
+  idx <- arg_match0(x, names(constants), arg_nm = "auth", error_call = caller_env())
+  constants[[]]
 }