+++ NEWS.md
@@ -1,5 +1,7 @@
 # httr2 (development version)
 
+* Added PKCE support to `oauth_flow_device()` (@flahn, #92).
+
 # httr2 0.1.1
 
 * Fix R CMD check failures on CRAN