+++ NEWS.md
@@ -1,5 +1,8 @@
 # httr2 (development version)
 
+* `httr_path` class renamed to `httr2_path` to correctly match package name 
+  (#99).
+
 * Added PKCE support to `oauth_flow_device()` (@flahn, #92).
 
 # httr2 0.1.1